+++ addons/website/static/tests/interactions/image_shape_hover_effect.test.js
@@ -26,7 +26,7 @@ test("image_shape_hover_effect changes image on enter & leave", async () => {
                 data-mimetype="image/svg+xml" data-original-id="276"
                 data-original-src="/website/static/src/img/snippets_demo/s_banner_3.jpg"
                 data-mimetype-before-conversion="image/jpeg"
-                data-shape="web_editor/geometric/geo_door" data-file-name="s_banner_3.svg"
+                data-shape="html_builder/geometric/geo_door" data-file-name="s_banner_3.svg"
                 data-shape-colors=";;;;" data-original-mimetype="image/jpeg"
                 data-x="160" data-y="0"
                 data-width="640" data-height="640"
