+++ addons/website/views/snippets/s_closer_look.xml
@@ -13,7 +13,7 @@
             <p>
                 <br/>
             </p>
-            <img class="img-fluid mx-auto" data-shape="web_editor/devices/macbook_front" data-file-name="s_closer_look.jpg" data-original-mimetype="image/jpeg" src="/web_editor/image_shape/website.s_closer_look_default_image/web_editor/devices/macbook_front.svg" alt=""/>
+            <img class="img-fluid mx-auto" data-shape="html_builder/devices/macbook_front" data-file-name="s_closer_look.jpg" data-original-mimetype="image/jpeg" src="/html_builder/image_shape/website.s_closer_look_default_image/html_builder/devices/macbook_front.svg" alt=""/>
         </div>
     </section>
 </template>
