+++ addons/website/views/snippets/s_company_team_shapes.xml
@@ -8,7 +8,7 @@
             <p><br/></p>
             <div class="row">
                 <div class="col-6 col-lg-4 pb24">
-                    <p><img src="/web_editor/image_shape/website.s_company_team_image_1/web_editor/geometric_round/geo_round_square.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric_round/geo_round_square" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_1.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
+                    <p><img src="/html_builder/image_shape/website.s_company_team_image_1/html_builder/geometric_round/geo_round_square.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric_round/geo_round_square" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_1.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
                     <h3 class="h5-fs" style="text-align: center;">Tony Fred</h3>
                 </div>
                 <div class="col-6 col-lg-4 pb24">
@@ -16,7 +16,7 @@
                     <h3 class="h5-fs" style="text-align: center;">Mich Stark</h3>
                 </div>
                 <div class="col-6 col-lg-4 pb24">
-                    <p><img src="/web_editor/image_shape/website.s_company_team_image_3/web_editor/geometric/geo_sonar.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric/geo_sonar" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_3.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
+                    <p><img src="/html_builder/image_shape/website.s_company_team_image_3/html_builder/geometric/geo_sonar.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric/geo_sonar" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_3.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
                     <h3 class="h5-fs" style="text-align: center;">Aline Turner</h3>
                 </div>
                 <div class="col-6 col-lg-4 pb24">
@@ -24,11 +24,11 @@
                     <h3 class="h5-fs" style="text-align: center;">Iris Joe</h3>
                 </div>
                 <div class="col-6 col-lg-4 pb24">
-                    <p><img src="/web_editor/image_shape/website.s_company_team_image_5/web_editor/geometric/geo_door.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric/geo_door" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_5.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
+                    <p><img src="/html_builder/image_shape/website.s_company_team_image_5/html_builder/geometric/geo_door.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric/geo_door" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_5.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
                     <h3 class="h5-fs" style="text-align: center;">Pete Bluestork</h3>
                 </div>
                 <div class="col-6 col-lg-4 pb24">
-                    <p><img src="/web_editor/image_shape/website.s_company_team_image_6/web_editor/geometric_round/geo_round_circle.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric_round/geo_round_circle" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_6.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
+                    <p><img src="/html_builder/image_shape/website.s_company_team_image_6/html_builder/geometric_round/geo_round_circle.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric_round/geo_round_circle" data-original-mimetype="image/jpeg" data-file-name="s_company_team_image_6.svg" data-shape-colors=";;;;" style="width: 100% !important;"/></p>
                     <h3 class="h5-fs" style="text-align: center;">Sophia Langston</h3>
                 </div>
             </div>
