+++ addons/website/views/snippets/s_cta_mobile.xml
@@ -12,7 +12,7 @@
                     <a href="https://play.google.com/store/"><img src="/web/image/website.google_play_image" class="img img-fluid" alt="Google Play"/></a>
                 </div>
                 <div class="o_grid_item o_grid_item_image o_snippet_mobile_invisible g-height-8 g-col-lg-4 col-lg-4 d-none d-lg-block" style="grid-area: 1 / 8 / 9 / 12; z-index: 2; --grid-item-padding-y: 0px; --grid-item-padding-x: 0px;">
-                    <img src="web_editor/image_shape/website.s_cta_mockups_default_image_1/web_editor/devices/galaxy_front_portrait_half.svg" class="img img-fluid" data-shape="web_editor/devices/galaxy_front_portrait_half" data-shape-colors=";;;;#111827" data-original-mimetype="image/jpeg" data-file-name="s_cta_mockups_1.jpg" alt=""/>
+                    <img src="html_builder/image_shape/website.s_cta_mockups_default_image_1/html_builder/devices/galaxy_front_portrait_half.svg" class="img img-fluid" data-shape="html_builder/devices/galaxy_front_portrait_half" data-shape-colors=";;;;#111827" data-original-mimetype="image/jpeg" data-file-name="s_cta_mockups_1.jpg" alt=""/>
                 </div>
             </div>
         </div>
