+++ addons/website/views/snippets/s_image_hexagonal.xml
@@ -6,7 +6,7 @@
         <div class="container-fluid">
             <div class="row o_grid_mode" data-row-count="14">
                 <div class="o_grid_item o_grid_item_image g-col-lg-4 g-height-9 col-lg-4" style="grid-area: 3 / 2 / 12 / 6; z-index: 1;">
-                    <img src="/web_editor/image_shape/website.s_image_hexagonal_default_image_1/web_editor/geometric/geo_hexagon.svg" class="img img-fluid" data-shape="web_editor/geometric/geo_hexagon" data-file-name="s_image_text.jpg" data-original-mimetype="image/jpeg" alt=""/>
+                    <img src="/html_builder/image_shape/website.s_image_hexagonal_default_image_1/html_builder/geometric/geo_hexagon.svg" class="img img-fluid" data-shape="html_builder/geometric/geo_hexagon" data-file-name="s_image_text.jpg" data-original-mimetype="image/jpeg" alt=""/>
                 </div>
                 <div class="o_grid_item g-col-lg-4 g-height-10 col-lg-4" style="grid-area: 3 / 5 / 13 / 9; z-index: 2;">
                     <h2 class="display-4-fs">Empowering<br/>Innovative<br/>Solutions</h2>
