+++ addons/website/views/snippets/s_images_constellation.xml
@@ -12,13 +12,13 @@
                     <img class="img img-fluid mx-auto rounded" src="/web/image/website.s_picture_default_image" alt=""/>
                 </div>
                 <div class="o_grid_item o_grid_item_image g-col-lg-4 g-height-6 col-lg-4 d-none d-lg-block o_snippet_mobile_invisible" style="z-index: 3; grid-area: 7 / 9 / 13 / 13; --grid-item-padding-x: 0px; --grid-item-padding-y: 24px;">
-                    <img class="img img-fluid mx-auto rounded" alt=""  data-shape="web_editor/geometric/geo_sonar" data-file-name="s_images_constellation_default_image.jpg" data-original-mimetype="image/jpeg/" src="/web_editor/image_shape/website.s_images_constellation_default_image/web_editor/geometric/geo_sonar.svg"/>
+                    <img class="img img-fluid mx-auto rounded" alt=""  data-shape="html_builder/geometric/geo_sonar" data-file-name="s_images_constellation_default_image.jpg" data-original-mimetype="image/jpeg/" src="/html_builder/image_shape/website.s_images_constellation_default_image/html_builder/geometric/geo_sonar.svg"/>
                 </div>
                 <div class="o_grid_item g-height-4 g-col-lg-6 col-12 col-lg-6" style="z-index: 6; grid-area: 12 / 5 / 16 / 11; --grid-item-padding-x: 16px; --grid-item-padding-y: 16px;">
                     <h2 style="text-align: right;">A constellation of amazing solutions tailored for your needs</h2>
                 </div>
                 <div class="o_grid_item o_grid_item_image g-col-lg-4 g-height-7 col-lg-4 d-none d-lg-block o_snippet_mobile_invisible" style="z-index: 4; grid-area: 15 / 2 / 22 / 6; --grid-item-padding-x: 0px; --grid-item-padding-y: 0px;">
-                    <img class="img img-fluid mx-auto rounded" alt="" data-shape="web_editor/composite/composite_double_pill" data-file-name="s_images_constellation_default_image_1.jpg" data-original-mimetype="image/jpeg/" src="/web_editor/image_shape/website.s_images_constellation_default_image_1/web_editor/composite/composite_double_pill.svg"/>
+                    <img class="img img-fluid mx-auto rounded" alt="" data-shape="html_builder/composite/composite_double_pill" data-file-name="s_images_constellation_default_image_1.jpg" data-original-mimetype="image/jpeg/" src="/html_builder/image_shape/website.s_images_constellation_default_image_1/html_builder/composite/composite_double_pill.svg"/>
                 </div>
                 <div class="o_grid_item o_grid_item_image g-col-lg-5 g-height-9 col-lg-5 d-none d-lg-block o_snippet_mobile_invisible" style="z-index: 5; grid-area: 15 / 8 / 24 / 13; --grid-item-padding-x: 0px; --grid-item-padding-y: 0px;">
                     <img class="img img-fluid mx-auto rounded" src="/web/image/website.s_images_constellation_default_image_2" alt=""/>
