+++ addons/website/views/snippets/s_images_mosaic.xml
@@ -10,19 +10,19 @@
                     <p class="lead">Discover our latest solutions for your business.</p>
                 </div>
                 <div class="col-6 col-lg-4 p-0">
-                    <img src="/web_editor/image_shape/website.s_images_mosaic_default_image_1/web_editor/geometric/geo_diamond.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric/geo_diamond" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_1.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
+                    <img src="/html_builder/image_shape/website.s_images_mosaic_default_image_1/html_builder/geometric/geo_diamond.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric/geo_diamond" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_1.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
                 </div>
                 <div class="col-6 col-lg-4 p-0">
-                    <img src="/web_editor/image_shape/website.s_images_mosaic_default_image_2/web_editor/geometric/geo_sonar.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric/geo_sonar" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_2.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
+                    <img src="/html_builder/image_shape/website.s_images_mosaic_default_image_2/html_builder/geometric/geo_sonar.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric/geo_sonar" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_2.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
                 </div>
                 <div class="col-6 col-lg-4 p-0">
-                    <img src="/web_editor/image_shape/website.s_images_mosaic_default_image_3/web_editor/geometric_round/geo_round_circle.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric_round/geo_round_circle" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_3.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
+                    <img src="/html_builder/image_shape/website.s_images_mosaic_default_image_3/html_builder/geometric_round/geo_round_circle.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric_round/geo_round_circle" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_3.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
                 </div>
                 <div class="col-6 col-lg-4 p-0">
-                    <img src="/web_editor/image_shape/website.s_images_mosaic_default_image_4/web_editor/geometric/geo_tear.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric/geo_tear" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_4.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
+                    <img src="/html_builder/image_shape/website.s_images_mosaic_default_image_4/html_builder/geometric/geo_tear.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric/geo_tear" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_4.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
                 </div>
                 <div class="col-6 col-lg-4 p-0">
-                    <img src="/web_editor/image_shape/website.s_images_mosaic_default_image_5/web_editor/geometric/geo_door.svg" class="img img-fluid" alt="" data-shape="web_editor/geometric/geo_door" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_5.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
+                    <img src="/html_builder/image_shape/website.s_images_mosaic_default_image_5/html_builder/geometric/geo_door.svg" class="img img-fluid" alt="" data-shape="html_builder/geometric/geo_door" data-original-mimetype="image/jpeg" data-file-name="s_images_mosaic_default_image_5.svg" data-shape-colors=";;;;" style="width: 100% !important;"/>
                 </div>
             </div>
         </div>
