+++ addons/website/views/snippets/s_intro_pill.xml
@@ -6,8 +6,8 @@
         <div class="container-fluid">
             <div class="row o_grid_mode" data-row-count="11">
                 <div class="o_grid_item oe_img_bg o_grid_item_image g-col-lg-5 g-height-9 col-lg-5 order-lg-0" style="order: 2; z-index: 1; grid-area: 1 / 1 / 10 / 6; --grid-item-padding-y: 0px; --grid-item-padding-x: 0px;">
-                    <img class="img-fluid mx-auto" data-shape="web_editor/geometric_round/geo_round_pill" data-file-name="s_intro_pill_default_image.jpg"
-                    data-original-mimetype="image/jpeg" src="/web_editor/image_shape/website.s_intro_pill_default_image/web_editor/geometric_round/geo_round_pill.svg" alt=""/>
+                    <img class="img-fluid mx-auto" data-shape="html_builder/geometric_round/geo_round_pill" data-file-name="s_intro_pill_default_image.jpg"
+                    data-original-mimetype="image/jpeg" src="/html_builder/image_shape/website.s_intro_pill_default_image/html_builder/geometric_round/geo_round_pill.svg" alt=""/>
                 </div>
                 <div class="o_grid_item g-height-11 g-col-lg-8 col-lg-8 order-lg-0" data-name="Box" style="order: 1; z-index: 3; grid-area: 1 / 3 / 12 / 11; --grid-item-padding-y: 120px;">
                     <h1 class="display-3" style="text-align: center;">Experience the<br/>best quality services</h1>
@@ -19,7 +19,7 @@
                     </p>
                 </div>
                 <div class="o_grid_item o_grid_item_image g-col-lg-5 g-height-10 d-none d-lg-block col-lg-5 order-lg-0" style="order: 3; z-index: 2; grid-area: 2 / 8 / 12 / 13; --grid-item-padding-y: 0px; --grid-item-padding-x: 0px;">
-                    <img class="img-fluid mx-auto" data-shape="web_editor/geometric_round/geo_round_pill" data-file-name="s_intro_pill_default_image_2.jpg" data-original-mimetype="image/jpeg" src="/web_editor/image_shape/website.s_intro_pill_default_image_2/web_editor/geometric_round/geo_round_pill.svg" alt=""/>
+                    <img class="img-fluid mx-auto" data-shape="html_builder/geometric_round/geo_round_pill" data-file-name="s_intro_pill_default_image_2.jpg" data-original-mimetype="image/jpeg" src="/html_builder/image_shape/website.s_intro_pill_default_image_2/html_builder/geometric_round/geo_round_pill.svg" alt=""/>
                 </div>
             </div>
         </div>
