+++ addons/website/views/snippets/s_mockup_image.xml
@@ -2,7 +2,7 @@
 <odoo>
 
 <template id="s_mockup_image" name="Mockup Image">
-    <section class="s_mockup_image o_cc o_cc1 pt72 pb72" data-oe-shape-data="{'shape':'web_editor/Wavy/11_001', 'showOnMobile':true}">
+    <section class="s_mockup_image o_cc o_cc1 pt72 pb72" data-oe-shape-data="{'shape':'html_builder/Wavy/11_001', 'showOnMobile':true}">
         <div class="o_we_shape o_web_editor_Wavy_11_001 o_shape_show_mobile"/>
         <div class="container">
             <div class="row align-items-center">
@@ -13,7 +13,7 @@
                     <p><a href="#" class="btn btn-primary o_translate_inline">Learn more</a></p>
                 </div>
                 <div class="col-lg-8 pt16 pb16">
-                    <img src="/web_editor/image_shape/website.s_mockup_image_default_image/web_editor/devices/macbook_front.svg" class="img-fluid ms-auto" data-shape="web_editor/devices/macbook_front" data-shape-colors=";;#F3F2F2;;" data-file-name="s_text_image.jpg" data-original-mimetype="image/jpeg" alt=""/>
+                    <img src="/html_builder/image_shape/website.s_mockup_image_default_image/html_builder/devices/macbook_front.svg" class="img-fluid ms-auto" data-shape="html_builder/devices/macbook_front" data-shape-colors=";;#F3F2F2;;" data-file-name="s_text_image.jpg" data-original-mimetype="image/jpeg" alt=""/>
                 </div>
             </div>
         </div>
