+++ addons/website/views/snippets/s_shape_image.xml
@@ -6,7 +6,7 @@
         <div class="container">
             <div class="row align-items-center">
                 <div class="col-lg-5 pt16 pb16">
-                    <img class="img img-fluid" data-shape="web_editor/composite/composite_double_pill" data-file-name="s_shape_image_default_image.jpg" data-original-mimetype="image/jpeg" src="/web_editor/image_shape/website.s_shape_image_default_image/web_editor/composite/composite_double_pill.svg" alt=""/>
+                    <img class="img img-fluid" data-shape="html_builder/composite/composite_double_pill" data-file-name="s_shape_image_default_image.jpg" data-original-mimetype="image/jpeg" src="/html_builder/image_shape/website.s_shape_image_default_image/html_builder/composite/composite_double_pill.svg" alt=""/>
                 </div>
                 <div class="col-lg-6 offset-lg-1 pt16 pb16">
                     <h2 class="h3-fs">About our product line</h2>
